+++ R/geom-label-pairwise.r
@@ -15,7 +15,7 @@ geom_label_pairwise <-
            parse = FALSE,
            nudge_x = 0,
            nudge_y = 0,
-           default.colour = "black",
+           default.colour = NULL,
            default.color = default.colour,
            colour.target = "all",
            color.target = colour.target,
@@ -85,24 +85,24 @@ GeomLabelPairwise <-
                    default_aes = ggplot2::aes(
                      x = NA_real_,
                      colour = "black",
-                     fill = "white",
-                     size = 3.88,
-                     angle = 0, # currently ignored
-                     linewidth = 0.5, # same as segment.linewidth default
-                     linetype = "solid",
+                     fill = rgb(1, 1, 1, alpha = 0.75), # "white", but occluded data are visible
+                     family = "",
+                     size = 3.87,
+                     angle = 0,
                      hjust = 0.5,
                      vjust = 0.5,
                      alpha = NA,
-                     family = "",
                      fontface = 1,
-                     lineheight = 1.2
+                     lineheight = 1.2,
+                     linewidth = 0.5,
+                     linetype  = "solid"
                    ),
 
                    draw_panel = function(data, panel_params, coord, #panel_scales,
                                          parse = FALSE,
                                          na.rm = FALSE,
                                          size.unit = "mm",
-                                         default.colour = "black",
+                                         default.colour = NULL,
                                          colour.target = "all",
                                          default.alpha = NA,
                                          alpha.target = "fill",
@@ -110,6 +110,8 @@ GeomLabelPairwise <-
                                          arrow = NULL,
                                          label.padding = unit(0.25, "lines"),
                                          label.r = unit(0.15, "lines")) {
+
+                     default.colour <- check_default_colour(default.colour)
 
                      data$label <- as.character(data$label)
                      data <- subset(data, !is.na(label) & label != "")