+++ tests/testthat/test-geom-quadrant-lines.R
@@ -7,6 +7,31 @@ library(ggplot2)
     group = c("A", "A", "A", "B", "B", "B")
   )
 
+test_that("geom_quadrant_lines errors on missing intercept", {
+  expect_error(
+    ggplot(tst.df, aes(x = x, y = y)) +
+      geom_point() +
+      geom_quadrant_lines(xintercept = NULL)
+  )
+
+  expect_error(
+    ggplot(tst.df, aes(x = x, y = y)) +
+      geom_point() +
+      geom_quadrant_lines(yintercept = NULL)
+  )
+
+  # errors in 'ggplot2' code!
+  # expect_error(
+  #   ggplot(tst.df, aes(x = x, y = y)) +
+  #     geom_point() +
+  #     geom_quadrant_lines(xintercept = NULL,
+  #                         yintercept = NULL)
+  # )
+
+})
+
+# tests with intercept coords mapped with aes() are missing!!
+
 test_that("geom_quadrant_lines works as expected with default values", {
   res <- ggplot(tst.df, aes(x = x, y = y)) +
     geom_point() +
@@ -22,9 +47,8 @@ test_that("geom_quadrant_lines works as expected with default values", {
     PANEL = factor(1),
     group = gp,
     colour = "black",
-    size = 0.5,
     linewidth = 0.5,
-    linetype = "dashed",
+    linetype = 1,
     alpha = NA
   )
 
@@ -51,7 +75,6 @@ test_that("geom_quadrant_lines works as expected with custom parameter values", 
     PANEL = factor(1),
     group = gp,
     colour = "blue",
-    size = 0.5,
     linewidth = 0.5,
     linetype = "dashed",
     alpha = NA
@@ -93,6 +116,19 @@ test_that("geom_vhlines returns an error if neither x nor y intercept is provide
         geom_vhlines()
     )
   )
+
+  expect_error(
+    ggplot(tst.df, aes(x = x, y = y)) +
+      geom_point() +
+      geom_vhlines(xintercept = 0)
+  )
+
+  expect_error(
+    ggplot(tst.df, aes(x = x, y = y)) +
+      geom_point() +
+      geom_vhlines(yintercept = 0)
+  )
+
 })
 
 test_that("geom_vhlines works as expected with multiple intercepts", {
@@ -113,7 +149,6 @@ test_that("geom_vhlines works as expected with multiple intercepts", {
     PANEL = factor(c(1, 1)),
     group = gp,
     colour = c("black", "black"),
-    size = c(0.5, 0.5),
     linewidth = c(0.5, 0.5),
     linetype = c(1, 1),
     alpha = c(NA, NA)